clean and stb introduction
