--- Raster/get_json_feature.py
@@ -1,5 +1,0 @@
-import json
-
-def getFeatures(gdf):
-    """Function to parse features from GeoDataFrame in such a manner that rasterio wants them"""
-    return [json.loads(gdf.to_json())['features'][0]['geometry']]
