--- Raster/index_exercise.txt
@@ -1,2 +1,0 @@
-Hello, I am test
-Me to
